requirements.txt
stix2
requests
geoip2
pyyaml
python-dotenv
pycti==6.4.2